Add SwissPlayers component and refactor
